Skal validere at bruker har samme ident som opprettet behandling
